requirements.txt
fastapi
uvicorn
python-dotenv
langgraph
langchain
playwright
google-generativeai
chromadb